translated imports and functions
